io500_fixed: verify/print if write phases are valid
For each of the write phases, check if it is valid, and mark the
phase result and aggregate score invalid if any are invalid.
Otherwise, the log messages can easily be missed.

For any write phase which is too short, provide an estimated size
parameter for that phase that should get the test run to be long
enough.  This requires adding a separate io500_ior_easy_size
parameter, so that the test can provide feedback on appropriate
sizes for this phase.  The actual units of io500_ior_easy_size are
arbitrary, only dependent on how io500_ior_easy_params uses it.

Signed-off-by: Andreas Dilger <adilger@dilger.ca>