find: download and build pfind by default
The pfind program is no longer in python, so it needs to be
downloaded and built to be usable.  There is no point in using
the serial find by default, since pfind will always be faster, and
that is just one more step needed by a user to get a good score.

Signed-off-by: Andreas Dilger <adilger@dilger.ca>