Update io500.sh
Under the parallel find section, we were still using the sfind.sh which usually doesn't find matches for the find. I tried on two systems and I had to change the pfind. So, the proposal is to use pfind by default. Moreover the binary of pfind was not declared to the variable.