ior: use random data seed for each IOR run
Instead of hard-coding the data verification seed (-G option) to
27 for each run, it should be generated randomly for each IOR run.
It needs to be the same for the write and read phases of ior-hard
and ior-easy, but does not need to be the same between the two.

Allow it to be set by the io500.sh script in case it is necessary
to run e.g. ior-hard-read multiple times to verify data generated
by a prior ior-hard-write run.

IOR only accepts a 31-bit seed, so limit random seed to avoid error.

Signed-off-by: Andreas Dilger <adilger@dilger.ca>
Closes #54