io500.sh: set tags for ISC20 runs and deactivate
Set the preferred 'io500-isc20' tags for the various Git repos.

Add a deprecation warning to the default io500.sh script so that
new users can clearly find it.  The message in io500_fixed.sh is
protected by $io500_ini, which is set in the new io500.sh script
in the io500-app repo, and allows this to be used as a subroutine.

Signed-off-by: Andreas Dilger <adilger@dilger.ca>